# [REVIEW] Performance related changes to get_dummies