--- a/python/cudf/cudf/dataframe/dataframe.py
+++ b/python/cudf/cudf/dataframe/dataframe.py
@@ -1569,6 +1569,7 @@ class DataFrame(object):
         3         4      bird          0          1.0          0.0          0.0
         4         5      fish          2          0.0          0.0          1.0
         """
+        cats = list(cats)
         newnames = [prefix_sep.join([prefix, str(cat)]) for cat in cats]
         newcols = self[column].one_hot_encoding(cats=cats, dtype=dtype)
         outdf = self.copy()