--- a/python/cudf/cudf/dataframe/dataframe.py
+++ b/python/cudf/cudf/dataframe/dataframe.py
@@ -1569,7 +1569,11 @@ class DataFrame(object):
         3         4      bird          0          1.0          0.0          0.0
         4         5      fish          2          0.0          0.0          1.0
         """
-        cats = list(cats)
+        if isinstance(cats, Series):
+            cats = cats.to_pandas()
+        else:
+            cats = pd.Series(cats)
+
         newnames = [prefix_sep.join([prefix, str(cat)]) for cat in cats]
         newcols = self[column].one_hot_encoding(cats=cats, dtype=dtype)
         outdf = self.copy()