<!--

Thank you for contributing to cuDF :)

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are being
   made.

2. Please ensure that you have written units tests for the changes made/features
   added.

3. If you are closing an issue please use one of the automatic closing words as
   noted here: https://help.github.com/articles/closing-issues-using-keywords/

4. If your pull request is not ready for review but you want to make use of the
   continuous integration testing facilities please label it with `[WIP]`.

5. If your pull request is ready to be reviewed without requiring additional
   work on top of it, then remove the `[WIP]` label (if present) and replace
   it with `[REVIEW]`. If assistance is required to complete the functionality,
   for example when the C/C++ code of a feature is complete but Python bindings
   are still required, then add the label `[HELP-REQ]` so that others can triage
   and assist. The additional changes then can be implemented on top of the
   same PR. If the assistance is done by members of the rapidsAI team, then no
   additional actions are required by the creator of the original PR for this,
   otherwise the original author of the PR needs to give permission to the
   person(s) assisting to commit to their personal fork of the project. If that
   doesn't happen then a new PR based on the code of the original PR can be
   opened by the person assisting, which then will be the PR that will be
   merged.

6. Once all work has been done and review has taken place please do not add
   features or make changes out of the scope of those requested by the reviewer
   (doing this just add delays as already reviewed code ends up having to be
   re-reviewed/it is hard to tell what is new etc!). Further, please do not
   rebase your branch on master/force push/rewrite history, doing any of these
   causes the context of any comments made by reviewers to be lost. If
   conflicts occur against master they should be resolved by merging master
   into the branch used for making the pull request.

Many thanks in advance for your cooperation!

-->

Fixes #2296 

Maybe we can hold on until #2202  gets merged so that typecast performance related changes I made can also go in to see if we have more to improve on performance side of this.

Changes:

- Previously `cats` was a device array and instead of iterating over it, bringing onto host and and passing it to `series.one_hot_encoding` as well where we iterate on `cats`. This changes results in following performance increase: 
previous performance:
~`7.82 ms ± 14.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)`~
`5.2 ms ± 9.26 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)`
new performance:
~`7.45 ms ± 11.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)`~
`4.77 ms ± 8.19 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)`